# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Al-Muhandis Islamic Intelligence Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /quran/surahs/{surahId}:
    get:
      operationId: getSurah
      tags: [quran]
      summary: Get a specific Surah
      parameters:
        - name: surahId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Surah details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Surah"
  /quran/surahs/{surahId}/ayahs:
    get:
      operationId: listAyahs
      tags: [quran]
      summary: Get all ayahs for a surah
      parameters:
        - name: surahId
          in: path
          required: true
          schema:
            type: integer
        - name: translationId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of ayahs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ayah"
  /quran/ayahs/{ayahId}:
    get:
      operationId: getAyah
      tags: [quran]
      summary: Get a specific ayah
      parameters:
        - name: ayahId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ayah details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ayah"
  /quran/ayahs/{ayahId}/words:
    get:
      operationId: getAyahWords
      tags: [quran]
      summary: Get word-by-word analysis for an ayah
      parameters:
        - name: ayahId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Word analysis
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AyahWord"
  /quran/ayahs/{ayahId}/tafseer:
    get:
      operationId: getAyahTafseer
      tags: [quran]
      summary: Get tafseer for an ayah
      parameters:
        - name: ayahId
          in: path
          required: true
          schema:
            type: integer
        - name: source
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Tafseer content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tafseer"
  /quran/ayahs/{ayahId}/related-hadiths:
    get:
      operationId: getAyahRelatedHadiths
      tags: [quran]
      summary: Get hadiths related to an ayah
      parameters:
        - name: ayahId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Related hadiths
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Hadith"
...
components:
  schemas:
...
    Surah:
      type: object
      properties:
        id:
          type: integer
        number:
          type: integer
        nameArabic:
          type: string
        nameTransliterated:
          type: string
        nameEnglish:
          type: string
        revelation:
          type: string
          enum: [Meccan, Medinan]
        ayahCount:
          type: integer
        juzNumber:
          type: integer
        description:
          type: string
      required: [id, number, nameArabic, nameTransliterated, nameEnglish, revelation, ayahCount, juzNumber]
    Ayah:
      type: object
      properties:
        id:
          type: integer
        surahId:
          type: integer
        surahName:
          type: string
        ayahNumber:
          type: integer
        arabicText:
          type: string
        translation:
          type: string
        transliteration:
          type: string
        juzNumber:
          type: integer
        pageNumber:
          type: integer
        audioUrl:
          type: string
          nullable: true
      required: [id, surahId, surahName, ayahNumber, arabicText, translation, transliteration, juzNumber, pageNumber]
    AyahWord:
      type: object
      properties:
        id:
          type: integer
        ayahId:
          type: integer
        position:
          type: integer
        arabicText:
          type: string
        transliteration:
          type: string
        translation:
          type: string
        rootWord:
          type: string
        grammar:
          type: string
      required: [id, ayahId, position, arabicText, transliteration, translation, rootWord, grammar]
    Tafseer:
      type: object
      properties:
        ayahId:
          type: integer
        source:
          type: string
        arabicText:
          type: string
        englishText:
          type: string
        scholarName:
          type: string
      required: [ayahId, source, arabicText, englishText, scholarName]
...
    Hadith:
      type: object
      properties:
        id:
          type: integer
        collectionId:
          type: string
        collectionName:
          type: string
        hadithNumber:
          type: string
        arabicText:
          type: string
        translation:
          type: string
        grade:
          type: string
          enum: [Sahih, Hasan, Da'if, Mawdu', Unknown]
        gradeReason:
          type: string
        narrator:
          type: string
        topics:
          type: array
          items:
            type: string
      required: [id, collectionId, collectionName, hadithNumber, arabicText, translation, grade, narrator, topics]
...